+++ examples/get_all.rs
@@ -1,3 +1,4 @@
+#[allow(clippy::needless_doctest_main)]
 fn main() {
     // Retrieve all listeners
     if let Ok(listeners) = listeners::get_all() {